pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Test models for BitTorrent-based Machine Learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"grpcio>=1.76.0",
	"numpy",
	"protobuf>=6.33.0",
	"torch>=2.9.0",
]

[dependency-groups]
dev = [
	"grpcio-tools>=1.76.0",
	"ruff",
]

[project.scripts]
model = "model.main:main"

[tool.pyright]
reportExplicitAny=false
reportAny=false

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"
# requires = ["setuptools", "setuptools-scm"]
# build-backend = "setuptools.build_meta"

[tool.uv.build-backend]
module-name = "model"
module-root = ""
source-exclude = [".git*", "Dockerfile", "Makefile", "*/data", "*/logs", "*/tools"]
wheel-exclude = [".git*", "Dockerfile", "Makefile", "*/data", "*/logs", "*/tools"]

[tool.setuptools]
packages = ["model", "model.lib.ipc"]
include-package-data = false

# [tool.setuptools.packages.find]
# where = ["."]
# include = ["model*"]
# namespaces = false

# [tool.setuptools.exclude-package-data]
# model = ["data", "tools", "Makefile", "Dockerfile"]
